Refactor EventSys a bit, upgrade typescript [Re: nmkataoka/adv-life#69]
Removed the eslint-plugin-deprecation package because typescript 4.0
now recognizes deprecation comments directly.
