+++ src/0-engine/ControllerList.ts
@@ -1,6 +1,6 @@
 import { CharacterCreationController } from '../2-backend-api/controllers/character-creation.controller';
 import { ShopController } from '../2-backend-api/controllers/shop-controller';
-import { Controller, ControllerConstructor } from './API/Controller';
+import { ControllerConstructor } from './API/Controller';
 
 const controllers: ControllerConstructor<any>[] = [CharacterCreationController, ShopController];
