+++ src/0-engine/ECS/event-system/event-callback.ts
@@ -1,0 +1,11 @@
+import { EntityManager } from '../EntityManager';
+
+export type AckCallback = (data: any) => void;
+
+export type EventCallbackArgs<Payload> = {
+  eMgr: EntityManager;
+  payload: Payload;
+  ack?: AckCallback;
+};
+
+export type EventCallback<Payload> = (args: EventCallbackArgs<Payload>) => void;
