+++ src/0-engine/SystemList.ts
@@ -3,7 +3,7 @@ import { AgentSys } from '../1-game-code/ecsystems/Agent/AgentSys';
 import { AttackSys } from '../1-game-code/ecsystems/AttackSys';
 import { ManaRegenSys } from '../1-game-code/ecsystems/ManaRegenSys';
 import { StatusEffectsSys } from '../1-game-code/ecsystems/Agent/StatusEffectsSys';
-import { EventSys } from './ECS/EventSys';
+import { EventSys } from './ECS/event-system/EventSys';
 import { CombatLogSys } from '../1-game-code/ecsystems/Combat/CombatLogSys';
 import { CharacterCreationController } from '../2-backend-api/controllers/character-creation.controller';
