+++ src/0-engine/index.ts
@@ -1,7 +1,7 @@
 export * from './ECS/built-in-components';
 export * from './API/Router';
 export { Controller } from './API/Controller';
-export type { EventCallback, EventCallbackArgs } from './ECS/EventSys';
+export type { EventCallback, EventCallbackArgs } from './ECS/event-system';
 export { ECSystem } from './ECS/ECSystem';
 export type {
   GetComponentFuncType,
