+++ src/1-game-code/ecsystems/Merchant/constants.ts
@@ -1,0 +1,1 @@
+export const BUY_ITEM_FROM_MERCHANT = 'merchant/buyItemFromMerchant';
