+++ src/1-game-code/ecsystems/Merchant/merchant.system.ts
@@ -1,6 +1,8 @@
 import { ECSystem, EventCallbackArgs } from '../../../0-engine';
 
-const entityBuysItemFromMerchant = (args: EventCallbackArgs<{ itemId: number }>) => {};
+const entityBuysItemFromMerchant = ({
+  payload,
+}: EventCallbackArgs<{ itemId: number; buyerInventoryId: number; sellerInventoryId: number }>) => {};
 
 export class MerchantSys extends ECSystem {
   public Start(): void {}
